+++ exports/light-editor.js
@@ -1,8 +1,6 @@
-import { html } from "lit";
-import HighlightJS from 'highlight.js/lib/core';
-import JavaScript from 'highlight.js/lib/languages/javascript';
-import HTML from 'highlight.js/lib/languages/xml';
-import CSS from 'highlight.js/lib/languages/css';
+import { html, render } from "lit";
+
+
 import { unsafeHTML } from "lit/directives/unsafe-html.js";
 import { ref } from "lit/directives/ref.js";
 
@@ -10,21 +8,31 @@ import { BaseElement } from "../internal/base-element.js";
 import { baseStyles } from "./base-styles.js";
 import { styles } from "./light-editor.styles.js";
 import { theme } from "./default-theme.styles.js";
-import { LightResizeEvent } from "./events/light-resize-event.js"
 import { dedent } from "../internal/dedent.js";
+import { LightResizeEvent } from "./events/light-resize-event.js";
+import { repeat } from "lit/directives/repeat.js";
+import { elementsToString } from "../internal/elements-to-strings.js";
+import { PrismHighlight, PrismJS } from "../internal/prism-highlight.js";
+import { LineNumberPlugin } from "../internal/line-number-plugin.js";
+import { Token } from "prismjs";
 
-HighlightJS.registerLanguage('javascript', JavaScript);
-HighlightJS.registerLanguage('xml', HTML);
-HighlightJS.registerLanguage('css', CSS);
+const newLineRegex = /\r\n?|\n/g
 
 /**
- * A bare bones plain text editor with syntax highlighting.
+ * A minimal plain text editor with syntax highlighting, line numbers, and line highlighting.
+ *    `<light-editor>` is not intended to replace full solutions like CodeMirror, but it
+ *    is a lightweight alternative using a `<pre>` overlaid on top of a `<textarea>`
+ *    and a few extra divs to make for an enjoyable text editing experience.
+ *
  * @customElement
  * @tagname light-editor
  *
+ * @event {Event} light-input - Re-emits the textarea's "input" event
  * @event {Event} light-change - Re-emits the textarea's "change" event
  * @event {Event} light-selectionchange - Re-emits the textarea's "selectionchange" event
- * @event {Event} light-input - Re-emits the textarea's "input" event
+ * @event {Event} light-focus - Re-emits the textarea's "focus" event
+ * @event {Event} light-blur - Re-emits the textarea's "blur" event
+ * @event {Event} light-resize - Is emitting whenever the editor resizes.
  *
  */
 export default class LightEditor extends BaseElement {
@@ -36,22 +44,24 @@ export default class LightEditor extends BaseElement {
     theme,
   ]
 
-  // One day.
+  // We will need to decide if we want to formAssociate or just mirror to a light DOM textarea.
   // static formAssociated = true
 
   static properties = {
-    label: {},
     value: {},
     language: {reflect: true},
+    hasInteracted: { type: Boolean, attribute: "has-interacted", reflect: true },
+    preserveWhitespace: { type: Boolean, reflect: true, attribute: "preserve-whitespace" }
   }
 
   constructor () {
     super()
 
     /**
+     * The language used for highlighting. Default is "html". "css" and "js" also included by default.
      * @type {string}
      */
-    this.language = 'xml'
+    this.language = 'html'
 
     /**
      * @type {string}
@@ -62,6 +72,134 @@ export default class LightEditor extends BaseElement {
      * @type {null | HTMLTextAreaElement}
      */
     this.textarea = null
+
+    /**
+     * Tracks if the user has interacted with the `<textarea>`
+     * @type {boolean}
+     */
+    this.hasInteracted = false
+
+    /**
+     * Whether to strip whitespace before first character, and after the last character.
+     * @type {boolean}
+     */
+    this.preserveWhitespace = false
+  }
+
+  /**
+   * @param {import("lit").PropertyValues<this>} changedProperties
+   */
+  willUpdate (changedProperties) {
+    if (this.value === this.getAttribute("value") && this.preserveWhitespace !== true) {
+      // Remove only lines that are blank with spaces that are blank. trim() removes preceding white-space for the line with characters.
+      // https://stackoverflow.com/questions/14572413/remove-line-breaks-from-start-and-end-of-string#comment104290392_14572494
+      this.value = dedent(this.value.replace(/(^\s*(?!.+)\n+)|(\n+\s+(?!.+)$)/g, "")).trim()
+    }
+
+    super.willUpdate(changedProperties)
+  }
+
+  render () {
+    const language = this.language
+
+    const highlightedCode = unsafeHTML(this.highlightCode({ code: this.value, language }))
+    this.syncScroll()
+
+    return html`
+			<div part="base">
+        <!-- Super important to not have white space here due to how white space is handled -->
+			  <div part="gutter"
+			    @focus=${() => {
+            this.textarea?.focus()
+			    }}
+			    @click=${(/** @type {Event} */ e) => {
+			      // Don't let a user focus on this area.
+            this.textarea?.focus()
+			    }}
+			  ></div>
+        <!-- This is where the fancy syntax highlighting comes in -->
+        <div part="base-editor">
+				  <pre
+            id="pre-${language}"
+            data-code-lang=${language}
+            aria-hidden="true"
+            part="pre pre-${language}"
+            class="language-${language}"
+            tabindex="-1"
+			      @focus=${() => {
+              this.textarea?.focus()
+			      }}
+			      @click=${(/** @type {Event} */ e) => {
+			        // Don't let a user focus on this area.
+              this.textarea?.focus()
+			      }}
+          ><code
+              part="code code-${language}"
+              class="language-${language}"
+            >${highlightedCode}</code></pre>
+          <!-- IMPORTANT! There must be no white-space above. -->
+				  <textarea
+				    aria-labelledby="label"
+            id="textarea-${language}"
+            data-code-lang=${language}
+            part="textarea textarea-${language}"
+            spellcheck="false"
+            autocorrect="off"
+            autocapitalize="off"
+            translate="no"
+            ${ref(this.textareaChanged)}
+            @keyup=${this.keyupHandler}
+            @keydown=${this.keydownHandler}
+            @focus=${() => {
+              this.hasInteracted = true
+              this.syncScroll()
+              this.setCurrentLineHighlight()
+              this.dispatchEvent(new Event("light-focus", { bubbles: true, composed: true }))
+            }}
+            @blur=${() => {
+              this.syncScroll()
+              this.setCurrentLineHighlight()
+              this.dispatchEvent(new Event("light-blur", { bubbles: true, composed: true }))
+            }}
+            @selectionchange=${/** @param {Event} e */ (e) => {
+              this.syncScroll()
+              this.setCurrentLineHighlight()
+              this.dispatchEvent(new Event("light-selectionchange", { bubbles: true, composed: true }))
+            }}
+            @click=${() => {
+              this.setCurrentLineHighlight()
+            }}
+            @pointerdown=${() => {
+              this.setCurrentLineHighlight()
+            }}
+            @pointerup=${() => {
+              this.setCurrentLineHighlight()
+            }}
+            @input=${/** @param {Event} e */ (e) => {
+              this.value = /** @type {HTMLTextAreaElement} */ (e.currentTarget).value
+              this.injectGutterCells()
+              this.setCurrentLineHighlight()
+              this.dispatchEvent(new Event("light-input", { bubbles: true, composed: true }))
+              this.syncScroll()
+            }}
+            @change=${/** @param {Event} e */ (e) => {
+              this.value = /** @type {HTMLTextAreaElement} */ (e.currentTarget).value
+              this.injectGutterCells()
+              this.setCurrentLineHighlight()
+              this.dispatchEvent(new Event("light-change", { bubbles: true, composed: true }))
+              this.syncScroll()
+            }}
+            @scroll=${/** @param {Event} e */ (e) => {
+              this.setCurrentLineHighlight()
+              this.syncScroll()
+            }}
+          >${this.value}</textarea>
+        </div> <!-- base-editor -->
+			</div> <!-- base -->
+
+      <!-- Hidden slot for holding content -->
+      <slot hidden @slotchange=${this.handleSlotChangeEvent}></slot>
+		`
   }
 
   /**
@@ -75,67 +213,52 @@ export default class LightEditor extends BaseElement {
       top, bottom
     } = entries[0].contentRect;
     const width = left + right
-    const height = top + bottom
-
-    // @ts-expect-error
-    target.parentElement.style.setProperty("--textarea-height", `${height}px`)
-
+    const height = top + bottom;
 
     /**
-     * Fires whenever the editor resizes
+     * Fires whenever the editor resizes, usually due to zoom in / out
      */
-    this.dispatchEvent(new LightResizeEvent("light-resize", {height, width}))
-    // One day we'll allow the textarea to resize the width.
-    // target.parentElement.style.setProperty("--textarea-width", `${width}px`)
+    this.dispatchEvent(new LightResizeEvent("light-resize", {height, width}));
+    this.syncScroll()
+    this.injectGutterCells()
   }
 
-  render () {
-    const language = this.language
+  /**
+   * @param {import("lit").PropertyValues<this>} changedProperties
+   */
+  updated (changedProperties) {
+    super.updated(changedProperties)
 
-    const highlightedCode = this.value ? unsafeHTML(this.highlightCode({ code: this.value, language })) : ""
+    this.syncScroll()
+    this.injectGutterCells()
+    setTimeout(() => this.setCurrentLineHighlight())
+  }
 
-    return html`
-			<div class="base" part="base">
-        <!-- This is where the fancy syntax highlighting comes in -->
-				<pre
-          id="pre-${language}"
-          data-code-lang=${language}
-          aria-hidden="true"
-          part="pre pre-${language}"
-        ><code
-            part="code code-${language}"
-            class="language-${language}"
-          >${highlightedCode}</code></pre>
-        <!-- IMPORTANT! There must be no white-space above. -->
-				<textarea
-          id="textarea-${language}"
-          data-code-lang=${language}
-          part="textarea textarea-${language}"
-          spellcheck="false"
-          autocorrect="off"
-          autocapitalize="off"
-          translate="no"
-          ${ref(this.textareaChanged)}
-          @keydown=${this.keydownHandler}
-          @selectionchange=${/** @param {Event} e */ (e) => {
-            this.dispatchEvent(new Event("light-selectionchange", { bubbles: true, composed: true }))
-          }}
-          @input=${/** @param {Event} e */ (e) => {
-            this.syncScroll(e)
-            this.value = /** @type {HTMLTextAreaElement} */ (e.currentTarget).value
-            this.dispatchEvent(new Event("light-input", { bubbles: true, composed: true }))
-          }}
-          @change=${/** @param {Event} e */ (e) => {
-            this.value = /** @type {HTMLTextAreaElement} */ (e.currentTarget).value
-            this.dispatchEvent(new Event("light-change", { bubbles: true, composed: true }))
-          }}
-          @scroll=${this.syncScroll}
-          .value=${this.value}
-        ></textarea>
-			</div>
+  /**
+   * Syncs the `<pre>` element scroll position to the same as the `<textarea>`
+   * @internal
+   */
+  syncScroll () {
+    /**
+     * @type {undefined | null | HTMLTextAreaElement}
+     */
+    const textarea = this.shadowRoot?.querySelector("textarea")
 
-      <slot hidden @slotchange=${this.handleSlotChangeEvent}></slot>
-		`
+    if (textarea == null) return
+
+    const pre = this.shadowRoot?.querySelector(`pre`)
+
+    if (pre) {
+      pre.scrollTop = textarea.scrollTop;
+      pre.scrollLeft = textarea.scrollLeft;
+    }
+
+    const gutter = this.shadowRoot?.querySelector("[part~='gutter']")
+
+    if (gutter) {
+      gutter.scrollTop = textarea.scrollTop;
+      gutter.scrollLeft = textarea.scrollLeft;
+    }
   }
 
   /**
@@ -148,22 +271,11 @@ export default class LightEditor extends BaseElement {
     }
 
     const textarea = element
+    this.textarea = textarea
 
     this.textareaResizeObserver = new ResizeObserver((entries) => this.handleTextAreaResize(entries))
-
     this.textareaResizeObserver.observe(textarea)
 
-    this.textareaMutationObserver = new MutationObserver((mutationRecords) => {
-      // We actually don't care about what the mutation is, just update and move on.
-      // for (const mutation of mutationRecords) {
-      // }
-      this.value = textarea.value
-    })
-
-    this.textareaMutationObserver.observe(textarea, {
-      characterData: true,
-      subtree: true
-    })
   }
 
   /**
@@ -180,40 +292,25 @@ export default class LightEditor extends BaseElement {
 
     const templates = slot.assignedElements({flatten: true})
 
-    const code = dedent(this.unescapeCharacters(templates.map((template) => template.innerHTML).join("\n")))
+    // We only unescape when passed into templates.
+    const code = dedent(this.unescapeTags(elementsToString(...templates)).trim())
 
-    if (code.trim()) {
+    if (code) {
       this.value = code
+      setTimeout(() => this.textarea?.setSelectionRange(0, 0))
       this.dispatchEvent(new Event("light-input", { bubbles: true, composed: true }))
       this.dispatchEvent(new Event("light-change", { bubbles: true, composed: true }))
     }
   }
 
   /**
-   * @internal
-   * @param {Event} e
+   * @ignore
+   * @param {KeyboardEvent} evt
    */
-  syncScroll (e) {
-    /**
-     * @type {null | HTMLTextAreaElement}
-     */
-    // @ts-expect-error
-    const textarea = e.target
-
-    if (textarea == null) return
-
-    const pre = this.shadowRoot?.querySelector(`pre`)
-
-    if (pre == null) return
-
-    pre.scrollTop = textarea.scrollTop;
-    pre.scrollLeft = textarea.scrollLeft;
-  }
-
-  disconnectedCallback () {
-    this.textareaResizeObserver?.disconnect()
-    this.textareaMutationObserver?.disconnect()
-    super.disconnectedCallback()
+  keyupHandler(evt) {
+    this.setCurrentLineHighlight()
+    // setTimeout is needed for Safari which appears to be "slow" to update selection APIs.
+    setTimeout(() => this.setCurrentLineHighlight())
   }
 
   /**
@@ -221,19 +318,47 @@ export default class LightEditor extends BaseElement {
    * @param {KeyboardEvent} evt
    */
   keydownHandler(evt) {
-    /**
-     * @type {HTMLTextAreaElement}
-     */
-    // @ts-expect-error
-    const target = evt.target
+    this.setCurrentLineHighlight()
+    // setTimeout is needed for Safari which appears to be "slow" to update selection APIs.
+    setTimeout(() => this.setCurrentLineHighlight())
+    // this.textarea
 
-    // Let's not focus trap. For now.
+    // Let's not trap focus. For now.
     // if ('Tab' === evt.key) {
     //   evt.preventDefault()
     //   target.setRangeText('\t', target.selectionStart, target.selectionEnd, 'end')
     //   this.value = target.value
     //   return
     // }
+  }
+
+  setCurrentLineHighlight () {
+    const code = this.shadowRoot?.querySelector("code")
+    const gutter = this.shadowRoot?.querySelector("[part~='gutter']")
+
+    const currentLineNumber = this.getCurrentLineNumber()
+
+    if (this.currentLineNumber === currentLineNumber) return
+
+    const prevLineNumber = this.currentLineNumber
+
+    // @ts-expect-error
+    this.currentLineNumber = currentLineNumber
+
+    if (currentLineNumber != null && currentLineNumber >= 0) {
+      const activeLineElement = code?.children[currentLineNumber]
+      const activeGutterElement = gutter?.children[currentLineNumber]
+
+      if (activeLineElement) {
+        code.children[prevLineNumber]?.classList?.remove("is-active")
+        activeLineElement.classList.add("is-active")
+      }
+
+      if (activeGutterElement) {
+        gutter.children[prevLineNumber]?.part?.remove("gutter-cell--active")
+        activeGutterElement.part.add("gutter-cell--active")
+      }
+    }
   }
 
   /**
@@ -244,21 +369,87 @@ export default class LightEditor extends BaseElement {
   highlightCode (options) {
     let { code, language } = options
 
-    code = this.unescapeCharacters(code)
-    // Dedent is nice, but we don't want to do it on user typed data.
-    // code = dedent(code)
     code = this.injectNewLine(code)
+    code = this.unescapeTags(code)
 
-    return HighlightJS.highlight(code, {language}).value
+    code = PrismHighlight(code, PrismJS.languages[language], language, {
+      afterTokenize: [
+        LineNumberPlugin(),
+        (env) => {
+          const currentToken = env.tokens[this.currentLineNumber]
+          if (!currentToken) return
+
+          if (currentToken instanceof Token) {
+            currentToken.type = currentToken.type + " is-active"
+          }
+        }
+      ]
+    })
+
+    /** We use this to wrap every line to perform line counting operations. */
+    return code
+  }
+
+  injectGutterCells () {
+    const gutter = this.shadowRoot?.querySelector("[part~='gutter']")
+
+    if (gutter) {
+      render(this.renderGutterCells(), /** @type {HTMLElement} */ (gutter))
+    }
+  }
+
+  getCurrentLineNumber () {
+    const textArea = this.textarea
+
+    if (!textArea) return
+
+    const textLines = textArea.value.substr(0, textArea.selectionStart).split(newLineRegex);
+    const currentLineNumber = textLines.length - 1;
+    // const currentColumnIndex = textLines[textLines.length-1].length;
+    // console.log("Current Line Number "+ currentLineNumber+" Current Column Index "+currentColumnIndex );
+
+    return currentLineNumber
+  }
+
+  renderGutterCells () {
+    const lines = this.shadowRoot?.querySelector("pre > code")?.children
+
+    if (!lines) return ``
+
+    const ary = Array.from(lines)
+
+    return repeat(ary, (el, index) => {
+      // @ts-expect-error
+      const height = /** @type {number} */ (el.offsetHeight)
+      const isCurrent = index === this.currentLineNumber
+
+      return index + height + isCurrent.toString()
+    }, (el ,index) => {
+      // @ts-expect-error
+      const height = /** @type {number} */ (el.offsetHeight)
+
+      const isCurrent = index === this.currentLineNumber
+
+      const parts = `gutter-cell ${isCurrent ? "gutter-cell--active" : ""}`
+      if (height) {
+        return html`<span part=${parts} style="${`height: ${height}px`}">${index}</span>`
+      }
+
+      return html`<span part=${parts}>${index}</span>`
+    })
   }
 
   /**
    * @ignore
    * @param {string} text
    */
-  unescapeCharacters (text) {
-    // Update code
-    return text.replaceAll("&lt;/script>", "</script>")
+  // This gets tricky. We could do this, but it may be unexpected...
+  unescapeTags (text) {
+    // Replace usages of `&lt;/script>` with `</script>`. Match against
+    // `&lt;/` so that other usages of &lt; aren't replaced.
+    // return text.replace(/&lt;\//g, '</');
+
+    return text.replaceAll(/&lt;\/([\w\d\.-_]+)>/g, "</$1>")
   }
 
   /**
@@ -268,11 +459,10 @@ export default class LightEditor extends BaseElement {
    */
   injectNewLine (text) {
     // Handle final newlines (see article)
-    if(text[text.length-1] == "\n") { // If the last character is a newline character
-      text += "\n"; // Add a placeholder space character to the final line
+    if(text === "" || text[text.length-1] === "\n") { // If the last character is a newline character
+      text += " "; // Add a placeholder space character to the final line
     }
 
     return text
   }
-
 }